Create the LFSPlugin class
Someone forced our hand on this, and so we now have the API in the
shape of pretty much a clone of "DB broker" aka DiskFile/DiskDir.
No idea if this is a good thing, but we need to move on, and this
works at least somewhat.

Of course the code in LFSPlugin runs on the context of Proxy server,
but otherwise it's a copy-paste job from corresponding storage nodes.

There is a unit test, but it's rather empty right now.