Get rid of split_path in LFSxxxxxController
The caller has already done the parsing, no reason to duplicate it.
Note that in the normal Swift, nodes call other nodes to update
summaries, so e.g. account nodes can be called with an additional
container argument. But in case of LFS this is unnecessary. We can
just invoke everything we need in a direct way.