Add DiskFile.usage and rearrange sections
Practice shows that people reading this document for the first time
get shocked by the pseudo-code section and cannot focus on the API
description, so let's relocate the pseudocode deeper into bowels.

The DiskFiel.usage is from review 30051, of course. We may be
able to persuade Peter to drop it from there.